Update to Processing 3 and improve overall stability
In addition to the minor changes necessary to update to Processing 3,
this also adds an interface element to show status, and makes the camera
loading much more robust. It also loads the camera in a separate thread
so that the UI pops up *much* faster with a message telling you what it
is doing.