Add basic video encoding support
This commit adds a new dependency on FFMPEG (included in the repository
for Windows and Linux 64-bit OSes). It uses the FFMPEG executable to
encode an mp4 version of the frames at 15 frames per second, including
an optional audio file in pretty much any audio format one would
normally encounter.