Refactor Waveform and add MultiRegionWaveform
Waveform has now been split into two classes, with Waveform itself
handling the drawing of the waveform and the motion of the cursor and
playhead, and a new subclass MultiRegionWaveform handling the selection
region. This will enable the Waveform class to be reused for sound
effects, where the length of the selection region can change.